Scope managment; draft albums

+++ server/dal/migrations/V2__chroma_scopes.sql
@@ -2,7 +2,7 @@ CREATE TABLE chroma_scopes (
     koala_id INT NOT NULL,
     scope VARCHAR(128) NOT NULL,
     granted_by INT NOT NULL,
-    granted_at INT NOT NULL,
+    granted_at BIGINT NOT NULL,
     PRIMARY KEY (koala_id, scope),
     FOREIGN KEY (koala_id) REFERENCES users(koala_id),
     FOREIGN KEY (granted_by) REFERENCES users(koala_id)
